Refactor cinema mode to a render prop component
