+++ src/Components/Video.js
@@ -1,25 +1,16 @@
-import styled, { injectGlobal } from 'styled-components'
-import is from 'styled-is'
+import styled from 'styled-components'
+
 import { Component } from 'preact'
-import { Col } from 'react-styled-flexboxgrid'
+
 import Flex from 'styled-flex-component'
 import { Link } from 'preact-router/match'
-import Portal from 'preact-portal'
+
 import LazyLoad from 'react-lazyload'
 import remcalc from 'remcalc'
 
 import Tag from './Styling/Tag'
 import Player from './Player'
-
-const Button = styled.button`
-  background: transparent;
-  display: block;
-  border: none;
-  color: #d62d22;
-  font-weight: bold;
-  text-align: right;
-  padding: 0;
-`
+import CinemaMode from './CinemaMode'
 
 const Speaker = styled.p`
   padding-left: ${remcalc(20)};
@@ -51,56 +42,11 @@ const Description = styled.p`
   line-height: ${remcalc(21)};
 `
 
-const Column = styled(Col)`
-  transition: all 200ms ease;
-  justify-content: center;
-  margin: 0 auto;
-  margin-bottom: ${remcalc(40)};
-
-  ${is('cinemaMode')`
-    position: fixed;
-    z-index: 9999;
-    top: 10%;
-    width: 90%;
-    left: 50%;
-    transform: translateX(-50%);
-    background: ${props => props.theme.white};
-    padding: ${remcalc(20)};
-    max-height: 90%;
-    overflow: scroll;
-    padding-bottom: ${remcalc(50)};
-  `};
-`
-
-const Overlay = styled.div`
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.9);
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 10;
-`
-
-injectGlobal`
-  body.cinema-mode {
-      overflow: hidden;
-  }
-`
-
 const makeLink = (url = 'speaker', name = 'FIX ME') =>
   `/${url}/${name.replace(/\s+/g, '-').toLowerCase()}`
 
-class VideoWarpper extends Component {
-  state = { cinemaMode: false, showVideo: false }
-
-  toggleCinemaMode = () => {
-    this.setState(({ cinemaMode }) => ({
-      cinemaMode: !cinemaMode,
-      showVideo: true
-    }))
-    document.body.classList.toggle('cinema-mode', this.state.cinemaMode)
-  }
+export class SimpleVideo extends Component {
+  state = { showVideo: false }
 
   showVideo = () => {
     this.setState(({ showVideo }) => ({
@@ -131,73 +77,73 @@ class VideoWarpper extends Component {
       removeFavorite,
       addFavorite,
       removeWatched,
-      addWatched
+      addWatched,
+      cinemaMode,
+      showCinemaVideo
     },
-    { cinemaMode, showVideo }
-  ) => (
-    <Column
-      cinemaMode={cinemaMode}
-      md={cinemaMode ? 12 : 4}
-      sm={cinemaMode ? 12 : 6}
-      xs={9}
-    >
-      <Player
-        showVideo={showVideo}
-        cinemaMode={cinemaMode}
-        id={id}
-        onClick={this.showVideo}
-        link={link}
-        name={name}
-        onEnd={() => this.endVideo(id)}
-      />
-      <Flex justifyBetween alignCenter>
-        <Name>{this.videoTitle(name)}</Name>
-        <Speaker>
-          {speaker.map(s => (
-            <Link
+    { showVideo }
+  ) => {
+    return (
+      <span>
+        <Player
+          showVideo={showVideo || showCinemaVideo}
+          cinemaMode={cinemaMode}
+          id={id}
+          onClick={this.showVideo}
+          link={link}
+          name={name}
+          onEnd={() => this.endVideo(id)}
+        />
+        <Flex justifyBetween alignCenter>
+          <Name>{this.videoTitle(name)}</Name>
+          <Speaker>
+            {speaker.map(s => (
+              <Link
+                key={s.id}
+                activeClassName="active"
+                href={makeLink('speaker', s.name)}
+              >
+                <span>{s.name}</span>
+              </Link>
+            ))}
+          </Speaker>
+        </Flex>
+        <Flex>
+          {tags.map(s => (
+            <Tag
               key={s.id}
               activeClassName="active"
-              href={makeLink('speaker', s.name)}
+              href={makeLink('category', s.name)}
             >
-              <span>{s.name}</span>
-            </Link>
+              #{s.name.toLowerCase()}
+            </Tag>
           ))}
-        </Speaker>
-      </Flex>
-      <Flex>
-        {tags.map(s => (
-          <Tag
-            key={s.id}
-            activeClassName="active"
-            href={makeLink('category', s.name)}
-          >
-            #{s.name.toLowerCase()}
-          </Tag>
-        ))}
-      </Flex>
-
-      <Button name="Toggle Cinema Mode" onClick={this.toggleCinemaMode}>
-        {cinemaMode ? 'Turn Off' : 'Turn On'} Cinema Mode
-      </Button>
-
-      {cinemaMode && description ? (
-        <Description>{description}</Description>
-      ) : null}
-
-      {cinemaMode ? (
-        <Portal into="body">
-          <Overlay onClick={this.toggleCinemaMode} />
-        </Portal>
-      ) : null}
-    </Column>
-  )
+        </Flex>
+        {cinemaMode && description ? (
+          <Description>{description}</Description>
+        ) : null}
+      </span>
+    )
+  }
 }
+
+const VideoWrapper = props => (
+  <CinemaMode
+    render={(cinemaMode, showCinemaVideo) => (
+      <SimpleVideo
+        {...props}
+        cinemaMode={cinemaMode}
+        showCinemaVideo={showCinemaVideo}
+      />
+    )}
+  />
+)
 
 export default ({ noLazy = false, talk }) =>
   noLazy ? (
-    <VideoWarpper {...talk} />
+    <VideoWrapper {...talk} />
   ) : (
     <LazyLoad height={310}>
-      <VideoWarpper {...talk} />
+      <VideoWrapper {...talk} />
     </LazyLoad>
   )
